+++ modules/AWSPowerShell/Cmdlets/Kendra/AWS.Tools.Kendra.Completers.psm1
@@ -80,6 +80,13 @@ $KNDR_Completers = {
 
     switch ($("$commandName/$parameterName"))
     {
+        # Amazon.Kendra.AttributeSuggestionsMode
+        "Update-KNDRQuerySuggestionsConfig/AttributeSuggestionsConfig_AttributeSuggestionsMode"
+        {
+            $v = "ACTIVE","INACTIVE"
+            break
+        }
+
         # Amazon.Kendra.ConditionOperator
         {
             ($_ -eq "New-KNDRDataSource/CustomDocumentEnrichmentConfiguration_PostExtractionHookConfiguration_InvocationCondition_Operator") -Or
@@ -196,6 +203,7 @@ $KNDR_Completers = {
 }
 
 $KNDR_map = @{
+    "AttributeSuggestionsConfig_AttributeSuggestionsMode"=@("Update-KNDRQuerySuggestionsConfig")
     "CustomDocumentEnrichmentConfiguration_PostExtractionHookConfiguration_InvocationCondition_Operator"=@("New-KNDRDataSource","Update-KNDRDataSource","Write-KNDRDocumentBatch")
     "CustomDocumentEnrichmentConfiguration_PreExtractionHookConfiguration_InvocationCondition_Operator"=@("New-KNDRDataSource","Update-KNDRDataSource","Write-KNDRDocumentBatch")
     "Edition"=@("New-KNDRIndex")